Move ListQueryOptions.ts to list/
