+++ packages/client/src/observable/internal/testUtils.ts
@@ -46,7 +46,7 @@ import type { ObjectPayload } from "../ObjectPayload.js";
 import type { OrderBy, Status } from "../ObservableClient/common.js";
 import type { Unsubscribable } from "../Unsubscribable.js";
 import type { Entry } from "./Layer.js";
-import type { ListQueryOptions } from "./ListQueryOptions.js";
+import type { ListQueryOptions } from "./list/ListQueryOptions.js";
 import type { ObjectCacheKey } from "./object/ObjectCacheKey.js";
 import type { OptimisticId } from "./OptimisticId.js";
 import type { Store } from "./Store.js";
